+++ src/create-layer.ts
@@ -1,18 +1,21 @@
 import type { Component } from 'vue'
-import type { LayerAdapter, LayerConfigStatic } from '@/types'
+import type { LayerConfigCreate } from '@/types'
 import { DEFAULT_CONTAINER_MODEL } from '@/types/config'
+import { mergeFragment } from '@/pipeline/merge-node-config'
 import { toFragmentFromStatic } from '@/pipeline/to-fragment'
 import { createUseLayer } from '@/instance/create-use-layer'
 
 export function createLayer(
   Container: Component,
-  config: LayerConfigStatic = {},
-  adapter?: LayerAdapter,
+  config: LayerConfigCreate = {},
 ) {
-  return createUseLayer({
-    Container,
-    create: toFragmentFromStatic(config),
-    defaultModel: config.model ?? DEFAULT_CONTAINER_MODEL,
-    adapter,
-  })
+  const { adapter, ...staticConfig } = config
+
+  const create = mergeFragment(
+    { container: { model: DEFAULT_CONTAINER_MODEL } },
+    toFragmentFromStatic(staticConfig),
+    { container: { component: Container } },
+  )
+
+  return createUseLayer({ create, adapter })
 }